refactor: separate notification requests from api.device to api.notifications
