+++ src/api/index.ts
@@ -4,6 +4,7 @@ import {achievements} from '@api/achievements';
 import {devices} from '@api/devices';
 import {inAppNotifications} from '@api/inAppNotifications';
 import {news} from '@api/news';
+import {notifications} from '@api/notifications';
 import {referrals} from '@api/referrals';
 import {statistics} from '@api/statistics';
 import {statusNotice} from '@api/statusNotice';
@@ -26,4 +27,5 @@ export const Api = Object.freeze({
   statusNotice,
   tasks,
   time,
+  notifications,
 });
