+++ src/api/notifications/index.ts
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: ice License 1.0
+
+import {getUserNotificationChannels} from './getUserNotificationChannels';
+import {toggleNotificationDomainOnOff} from './toggleNotificationDomainOnOff';
+
+export const notifications = Object.freeze({
+  getUserNotificationChannels,
+  toggleNotificationDomainOnOff,
+});
