+++ src/api/notifications/toggleNotificationDomainOnOff.ts
